refactor: small move to break dependency cycle
Will still review better option but unblock tests
