+++ openeogeotrellis/integrations/s3/__init__.py
@@ -1,3 +1,2 @@
 from openeogeotrellis.integrations.s3.bucket_details import BucketDetails, is_workspace_bucket
 from openeogeotrellis.integrations.s3.presigned_url import create_presigned_url
-from openeogeotrellis.integrations.s3.client import get_s3_client
