+++ openeogeotrellis/integrations/s3_client/__init__.py
@@ -1,0 +1,1 @@
+from openeogeotrellis.integrations.s3_client.client import get_s3_client
